tests/wpt/web-platform-tests/css/css-flexbox/parsing/flex-basis-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing flex-basis with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-basis">
<meta name="assert" content="flex-basis supports only the grammar 'content | <‘width’>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flex-basis", "none");
test_invalid_value("flex-basis", "auto content");
test_invalid_value("flex-basis", "-1px");
test_invalid_value("flex-basis", "-2%");
test_invalid_value("flex-basis", "3px 4%");
</script>
</body>
</html>
